support_ticket_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-ticket-env"
version = "0.1.0"
description = "Support Ticket Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m support_ticket_env.server.app
server = "support_ticket_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["support_ticket_env", "support_ticket_env.server"]
package-dir = { "support_ticket_env" = ".", "support_ticket_env.server" = "server" }